Fix flaky tests when re-write env
