+++ mqtt/mqtt-edgehub/Cargo.toml
@@ -31,6 +31,7 @@ lazy_static = "1.4"
 matches = "0.1"
 mockito = "0.25"
 proptest = "0.9"
+serial_test = "0.4"
 test-case = "1.0"
 tokio = { version = "0.2", features = ["macros"] }
