+++ mqtt/mqtt-edgehub/src/settings.rs
@@ -198,11 +198,14 @@ impl AuthConfig {
 
 #[cfg(test)]
 mod tests {
-    use std::{env, time::Duration};
+    use std::time::Duration;
+
+    use serial_test::serial;
 
     use mqtt_broker::settings::{
         BrokerConfig, HumanSize, QueueFullAction, RetainedMessagesConfig, SessionConfig,
     };
+    use mqtt_broker_tests_util::env;
 
     use super::{
         AuthConfig, CertificateConfig, ListenerConfig, Settings, TcpTransportConfig,
@@ -248,11 +251,13 @@ mod tests {
     }
 
     #[test]
+    #[serial(env_settings)]
     fn new_overrides_settings_from_env() {
         it_overrides_settings_from_env(Settings::new);
     }
 
     #[test]
+    #[serial(env_settings)]
     fn from_file_overrides_settings_from_env() {
         it_overrides_settings_from_env(|| Settings::from_file("config/default.json"));
     }
@@ -261,11 +266,11 @@ mod tests {
     where
         F: FnOnce() -> Result<Settings, ConfigError>,
     {
-        env::set_var("LISTENER__TCP__ADDRESS", "10.0.0.1:1883");
-        env::set_var("LISTENER__TLS__ADDRESS", "10.0.0.1:8883");
-        env::set_var("LISTENER__TLS__CERTIFICATE", "/tmp/edgehub/cert.pem");
-        env::set_var("LISTENER__TLS__PRIVATE_KEY", "/tmp/edgehub/pkey.pem");
-        env::set_var("AUTH__BASE_URL", "/auth/");
+        let _tcp_address = env::set_var("LISTENER__TCP__ADDRESS", "10.0.0.1:1883");
+        let _tls_address = env::set_var("LISTENER__TLS__ADDRESS", "10.0.0.1:8883");
+        let _tls_certificate = env::set_var("LISTENER__TLS__CERTIFICATE", "/tmp/edgehub/cert.pem");
+        let _tls_private_key = env::set_var("LISTENER__TLS__PRIVATE_KEY", "/tmp/edgehub/pkey.pem");
+        let _auth_base_url = env::set_var("AUTH__BASE_URL", "/auth/");
 
         let settings = make_settings().unwrap();
 
@@ -285,13 +290,14 @@ mod tests {
     }
 
     #[test]
+    #[serial(env_settings)]
     fn it_can_disable_default_options() {
         let settings = Settings::new().unwrap();
         assert!(settings.listener().tcp().is_some());
         assert!(settings.listener().tls().is_some());
 
-        env::set_var("LISTENER__TCP__ENABLED", "false");
-        env::set_var("LISTENER__TLS__ENABLED", "false");
+        let _tcp_enabled = env::set_var("LISTENER__TCP__ENABLED", "false");
+        let _tls_enabled = env::set_var("LISTENER__TLS__ENABLED", "false");
 
         let settings = Settings::new().unwrap();
         assert!(!settings.listener().tcp().is_some());
